+++ firmware/src/sensor/ma7xx.h
@@ -41,7 +41,7 @@ typedef struct
     Sensor base;
     MA7xxSensorConfig config;
     uint8_t errors;
-	int16_t angle;
+	int32_t angle;
 } MA7xxSensor;
 
 bool ma7xx_init_with_port(Sensor *s, const SSP_TYPE port, PAC55XX_SSP_TYPEDEF *ssp_struct);
@@ -65,7 +65,7 @@ static inline void ma7xx_send_angle_cmd(const Sensor *s)
 	ssp_write_one(((const MA7xxSensor *)s)->config.ssp_struct, MA_CMD_ANGLE);
 }
 
-static inline int16_t ma7xx_get_raw_angle(const Sensor *s)
+static inline int32_t ma7xx_get_raw_angle(const Sensor *s)
 {
     return ((const MA7xxSensor *)s)->angle;
 }
@@ -73,11 +73,11 @@ static inline int16_t ma7xx_get_raw_angle(const Sensor *s)
 static inline void ma7xx_update(Sensor *s, bool check_error)
 {
     MA7xxSensor *ms = (MA7xxSensor *)s;
-    const int16_t angle = ssp_read_one(ms->config.ssp_struct);
+    const int32_t angle = ssp_read_one(ms->config.ssp_struct);
 
     if (check_error)
     {
-    	const int16_t delta = ms->angle - angle;
+    	const int32_t delta = ms->angle - angle;
 		if ( ((delta > MAX_ALLOWED_DELTA) || (delta < -MAX_ALLOWED_DELTA)) &&
 		     ((delta > MAX_ALLOWED_DELTA_ADD) || (delta < MIN_ALLOWED_DELTA_ADD)) &&
 		     ((delta > MAX_ALLOWED_DELTA_SUB) || (delta < MIN_ALLOWED_DELTA_SUB)) )
